% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richardreeveDiversity-package.R
\docType{package}
\name{richardreeveDiversity-package}
\alias{richardreeveDiversity-package}
\alias{richardreeveDiversity}
\title{richardreeveDiversity R package}
\description{
A package that calculates the diversity of a bunch of communities derived
from the Barro Colorado Island forest census plot.
}
\author{
Richard Reeve <richard.reeve@glasgow.ac.uk>
}
